Sets sane defaults for number of cpus and memory size.
VMware UIs are the one assuring sane defaults when setting number of
vcpus and memory size. We need to make sure we apply the same defaults
in order to avoid weird behaviors or unexplainable issues.

This commit also adds two more functions to get the display name
assigned to the VM as well as any annotations the user may have written
down for the VM.